pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["evasion*"]

[project]
name = "evasion"
version = "0.1.0"
description = "Adversarial attack and explanation stability sweeps."
requires-python = ">=3.9"
dependencies = [
    "torch",
    "torchattacks",
    "captum",
    "numpy",
    "scipy",
    "pandas",
    "pyyaml",
]

[project.scripts]
evasion = "evasion.cli:main"